Feature: Server Performance improvements
Removed unused main methods, now only external libraries have main methods. Pretty pointless to keep these (cleanup!), Removed more unused classes, code blocks, such..

Improved server performance once again. now ~40.9mb startup idle. Rests at ~44.65mb fulltime now

Lots of changes made in this commit!

+++ content/com/rs/game/player/content/ItemSets.java
@@ -131,7 +131,7 @@ public class ItemSets {
 			player.getPackets().sendGameMessage("This isn't a set item.");
 			return;
 		}
-		String message = ClientScriptMap.getMap(1088).getStringValue(id);
+		String message = ClientScriptMap.getMap((short) 1088).getStringValue(id);
 		if (message == null)
 			return;
 		player.getPackets().sendGameMessage(message);
